+++ packages/generator/generators/module-generator/src/locale/en.ts
@@ -4,7 +4,7 @@ export const EN_LOCALE = {
   success: `Success！
 {packageManager} dev          # Run and debug the project
 {packageManager} build        # Generate build products in accordance with community norms and best practices
-{packageManager} lint:error   # Check and fix all codes
+{packageManager} lint         # Check and fix all codes
 {packageManager} change       # Add changeset for release changelog
 {packageManager} bump         # Generate revisions related to the release, such as updating the version number and generating changelog
 {packageManager} new          # Create more project elements`,