+++ packages/server/bff-utils/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @modern-js/bff-utils
 
+## 1.0.0-rc.7
+
+### Patch Changes
+
+- 224f7fe: fix server route match
+- 204c626: feat: initial
+- Updated dependencies [224f7fe]
+- Updated dependencies [204c626]
+  - @modern-js/bff-runtime@1.0.0-rc.7
+  - @modern-js/create-request@1.0.0-rc.7
+  - @modern-js/plugin@1.0.0-rc.7
+  - @modern-js/utils@1.0.0-rc.7
+
+## 1.0.0-rc.6
+
+### Patch Changes
+
+- 224f7fe: fix server route match
+- 204c626: feat: initial
+- Updated dependencies [224f7fe]
+- Updated dependencies [204c626]
+  - @modern-js/bff-runtime@1.0.0-rc.6
+  - @modern-js/create-request@1.0.0-rc.6
+  - @modern-js/plugin@1.0.0-rc.6
+  - @modern-js/utils@1.0.0-rc.6
+
+## 1.0.0-rc.5
+
+### Patch Changes
+
+- 224f7fe: fix server route match
+- 204c626: feat: initial
+- Updated dependencies [224f7fe]
+- Updated dependencies [204c626]
+  - @modern-js/bff-runtime@1.0.0-rc.5
+  - @modern-js/create-request@1.0.0-rc.5
+  - @modern-js/plugin@1.0.0-rc.5
+  - @modern-js/utils@1.0.0-rc.5
+
 ## 1.0.0-rc.4
 
 ### Patch Changes