+++ packages/server/plugin/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @modern-js/server-plugin
 
+## 1.0.0-rc.7
+
+### Patch Changes
+
+- 224f7fe: fix server route match
+- 204c626: feat: initial
+- Updated dependencies [224f7fe]
+- Updated dependencies [204c626]
+  - @modern-js/plugin@1.0.0-rc.7
+
+## 1.0.0-rc.6
+
+### Patch Changes
+
+- 224f7fe: fix server route match
+- 204c626: feat: initial
+- Updated dependencies [224f7fe]
+- Updated dependencies [204c626]
+  - @modern-js/plugin@1.0.0-rc.6
+
+## 1.0.0-rc.5
+
+### Patch Changes
+
+- 224f7fe: fix server route match
+- 204c626: feat: initial
+- Updated dependencies [224f7fe]
+- Updated dependencies [204c626]
+  - @modern-js/plugin@1.0.0-rc.5
+
 ## 1.0.0-rc.4
 
 ### Patch Changes